update in mem database

+++ src/components/database/in_memmory.rs
@@ -2,10 +2,18 @@ use super::DatabaseTrait;
 use super::RawEntity;
 use crate::common::BlockPtr;
 use crate::errors::DatabaseError;
+use crate::messages::EntityID;
+use crate::messages::EntityType;
 use crate::runtime::asc::native_types::store::Value;
 use std::collections::HashMap;
 
-pub type InMemoryDataStore = HashMap<String, HashMap<String, HashMap<String, Value>>>;
+type BlockPtrNumber = u64;
+type BlockPtrHash = String;
+type DeletedAt = Option<u64>;
+type EntityPayload = HashMap<String, Value>;
+type EntitySnapshots = Vec<(BlockPtrNumber, BlockPtrHash, DeletedAt, EntityPayload)>;
+
+pub type InMemoryDataStore = HashMap<EntityType, HashMap<EntityID, EntitySnapshots>>;
 
 impl DatabaseTrait for InMemoryDataStore {
     fn handle_load(
@@ -36,8 +44,8 @@ impl DatabaseTrait for InMemoryDataStore {
             return Ok(None);
         }
 
-        let entity = entity.unwrap().to_owned();
-        Ok(Some(entity))
+        let (_, _, _, data) = entity.unwrap().last().cloned().unwrap();
+        return Ok(Some(data));
     }
 
     fn handle_create(
@@ -53,38 +61,65 @@ impl DatabaseTrait for InMemoryDataStore {
 
         let table = store.get_mut(&entity_type).unwrap();
         if let Value::String(entity_id) = data.get("id").ok_or(DatabaseError::MissingID)? {
-            table.insert(entity_id.to_owned(), data);
+            // Check if this id exists or not
+            if table.get(entity_id).is_none() {
+                table.insert(entity_id.to_owned(), vec![]);
+            };
+
+            // Push new record
+            let mut snapshots = table.get_mut(entity_id).unwrap();
+            snapshots.push((block_ptr.number, block_ptr.hash, None, data));
+
             Ok(())
         } else {
             unimplemented!()
         }
     }
 
-    fn handle_update(
-        &mut self,
+    fn soft_delete(
+        &self,
         block_ptr: BlockPtr,
         entity_type: String,
         entity_id: String,
-        data: RawEntity,
-    ) -> Result<(), DatabaseError> {
-        self.handle_update_latest(entity_type, entity_id, data)
-    }
-
-    fn handle_update_latest(
-        &mut self,
-        entity_type: String,
-        entity_id: String,
-        data: RawEntity,
     ) -> Result<(), DatabaseError> {
         let store = self;
-        if !store.contains_key(&entity_type) {
-            store.insert(entity_type.clone(), HashMap::new());
-        }
-        assert!(data.contains_key("id"));
+        let table = store.get(&entity_type);
 
-        let table = store.get_mut(&entity_type).unwrap();
+        if table.is_none() {
+            return Err(DatabaseError::EntityTypeNotExists(entity_type));
+        }
+
+        let table = table.unwrap();
+        let entity = table.get_mut(&entity_id);
+
+        if entity.is_none() {
+            return Err(DatabaseError::EntityIDNotExists(entity_type, entity_id));
+        }
+
+        let mut snapshots = entity.unwrap();
+        for snapshot in snapshots.iter_mut() {
+            snapshot.2 = Some(block_ptr.number);
+        }
+
+        Ok(())
+    }
+
+    fn hard_delete(&self, entity_type: String, entity_id: String) -> Result<(), DatabaseError> {
+        let store = self;
+        let table = store.get(&entity_type);
+
+        if table.is_none() {
+            return Err(DatabaseError::EntityTypeNotExists(entity_type));
+        }
+
+        let table = table.unwrap();
+        let entity = table.get_mut(&entity_id);
+
+        if entity.is_none() {
+            return Err(DatabaseError::EntityIDNotExists(entity_type, entity_id));
+        }
+
         table.remove_entry(&entity_id);
-        table.insert(entity_id, data);
 
         Ok(())
     }
